socketbuffer.cxx, socket-unix.cxx, socket-win32.cxx: Fix size_t/unsigned int conversion warnings.

+++ log4cplus/src/socket-win32.cxx
@@ -271,7 +271,7 @@ log4cplus::helpers::getHostname (bool fqdn)
 
     while (true)
     {
-        ret = ::gethostname (&hn[0], hn.size () - 1);
+        ret = ::gethostname (&hn[0], static_cast<int>(hn.size ()) - 1);
         if (ret == 0)
         {
             hostname = &hn[0];
